+++ pareg_core/src/arg_into.rs
@@ -1,4 +1,11 @@
-use crate::{by_ref::ByRef, err::Result, from_arg::FromArg};
+use std::{
+    borrow::Cow,
+    ffi::{OsStr, OsString},
+    rc::Rc,
+    sync::Arc,
+};
+
+use crate::{err::Result, from_arg::FromArg, impl_all::impl_all};
 
 /// This trait represents a string reference object that can be parsed into a
 /// type.
@@ -12,20 +19,33 @@ pub trait ArgInto<'a> {
     /// assert_eq!("hello", "hello".arg_into::<&str>().unwrap());
     /// assert_eq!(5, "5".arg_into::<i32>().unwrap());
     /// ```
-    fn arg_into<T>(self) -> Result<T>
+    fn arg_into<T>(&'a self) -> Result<T>
     where
         T: FromArg<'a>;
 }
 
-impl<'a, S> ArgInto<'a> for S
-where
-    S: ByRef<&'a str>,
-{
-    #[inline(always)]
-    fn arg_into<T>(self) -> Result<T>
-    where
-        T: FromArg<'a>,
-    {
-        T::from_arg(self.by_ref())
+impl_all! { impl<'a> ArgInto<'a>:
+    OsStr, OsString, Arc<OsStr>, Rc<OsStr>, Cow<'a, OsStr>, &OsStr => {
+        #[inline(always)]
+        fn arg_into<T>(&'a self) -> Result<T>
+        where
+            T: FromArg<'a>,
+        {
+            #[allow(clippy::useless_asref)]
+            T::from_os_arg((*self).as_ref())
+        }
+    }
+}
+
+impl_all! { impl<'a> ArgInto<'a>:
+    str, String, Arc<str>, Rc<str>, Cow<'a, str>, &str => {
+        #[inline(always)]
+        fn arg_into<T>(&'a self) -> Result<T>
+        where
+            T: FromArg<'a>,
+        {
+            #[allow(clippy::useless_asref)]
+            T::from_arg((*self).as_ref())
+        }
     }
 }