+++ pareg_core/src/from_args.rs
@@ -1,7 +1,7 @@
-use crate::{ParegRef, Result};
+use crate::{ArgInto, ParegRef, Result};
 
 /// Macro for types that can parse command line arguments.
-pub trait FromArgs<'a, S: AsRef<str> = String>: Sized {
+pub trait FromArgs<'a>: Sized {
     /// Parse the given command line arguments into this type.
-    fn parse_args(args: &mut ParegRef<'a, S>) -> Result<Self>;
+    fn parse_args<S: ArgInto<'a>>(args: &mut ParegRef<'a, S>) -> Result<Self>;
 }