+++ pareg_core/src/impl_all.rs
@@ -1,9 +1,12 @@
 /// Bulk implementation of a trait.
 macro_rules! impl_all {
+    (impl<$lt:lifetime>: $($t:ty),* $(,)? => $body:tt) => {
+        $(impl<$lt> $t $body)*
+    };
     (impl<$lt:lifetime> $tr:ty: $($t:ty),* $(,)? => $body:tt) => {
         $(impl<$lt> $tr for $t $body)*
     };
-    ($tr:ty: $($t:ty),* $(,)? => $body:tt) => {
+    (impl $tr:ty: $($t:ty),* $(,)? => $body:tt) => {
         $(impl $tr for $t $body)*
     };
 }