+++ pareg_core/src/lib.rs
@@ -1,10 +1,10 @@
 mod arg_into;
-mod by_ref;
 pub mod check;
 mod err;
 mod from_arg;
 mod from_args;
 pub(crate) mod impl_all;
+mod lossy_string;
 mod pareg_ref;
 mod parsef;
 mod parsers;
@@ -15,17 +15,17 @@ mod utils;
 
 pub use crate::{
     arg_into::*,
-    by_ref::*,
     err::*,
     from_arg::*,
     from_args::*,
+    lossy_string::*,
     pareg_ref::*,
     parsef::*,
     parsers::*,
     reader::{FromRead, ReadFmt, Reader, ReaderChars, SetFromRead},
 };
 
-use std::{borrow::Cow, cell::Cell, env, ops::RangeBounds};
+use std::{borrow::Cow, cell::Cell, env, ffi::OsString, ops::RangeBounds};
 
 /// Helper for parsing arguments.
 ///
@@ -36,13 +36,13 @@ use std::{borrow::Cow, cell::Cell, env, ops::RangeBounds};
 /// structure by calling [`Pareg::ref_mut`] to pass around and do the parsing,
 /// because it can be less strict about lifetimes since it refers to the
 /// original pareg structure and so it is more powerful.
-pub struct Pareg {
-    args: Vec<String>,
+pub struct Pareg<S = String> {
+    args: Vec<S>,
     cur: Cell<usize>,
 }
 
-impl From<Vec<String>> for Pareg {
-    fn from(value: Vec<String>) -> Self {
+impl<S> From<Vec<S>> for Pareg<S> {
+    fn from(value: Vec<S>) -> Self {
         Self {
             args: value,
             cur: 0.into(),
@@ -50,34 +50,31 @@ impl From<Vec<String>> for Pareg {
     }
 }
 
-impl Pareg {
+impl<S> Pareg<S> {
     /// Create [`Pareg`] from vector of arguments. The first argument is NOT
     /// skipped.
     #[inline]
-    pub fn new(args: Vec<String>) -> Self {
+    pub fn new(args: Vec<S>) -> Self {
         args.into()
-    }
-
-    /// Create [`Pareg`] from [`env::args`], the first argument is skipped.
-    #[inline]
-    pub fn args() -> Self {
-        Self {
-            args: env::args().collect(),
-            cur: 1.into(),
-        }
     }
 
     /// DO NOT MAKE THIS PIBLIC. This can be public only if the lifetime
     /// captured inside [`ParegRef`] borrows the original [`Pareg`] mutably.
     #[inline(always)]
-    pub(crate) fn inner(&self) -> ParegRef<'_> {
+    pub(crate) fn inner<'a>(&'a self) -> ParegRef<'a, S>
+    where
+        S: ArgInto<'a>,
+    {
         ParegRef::new(&self.args, Cow::Borrowed(&self.cur))
     }
 
     /// Gets mutable reference to self. Mutating the resulting pareg ref will
     /// also mutate this pareg.
     #[inline]
-    pub fn get_mut_ref(&mut self) -> ParegRef<'_> {
+    pub fn get_mut_ref<'a>(&'a mut self) -> ParegRef<'a, S>
+    where
+        S: ArgInto<'a>,
+    {
         // It is OK to pass the inner reference out, because this will borrow
         // [`Pareg`] mutably and so the captured reference in [`ParegRef`]
         // also borrows [`Pareg`] mutably.
@@ -86,88 +83,144 @@ impl Pareg {
 
     /// Gets immutable reference to self. Mutating the resulting pareg ref will
     /// not mutate this pareg.
-    pub fn get_ref(&self) -> ParegRef<'_> {
+    pub fn get_ref<'a>(&'a self) -> ParegRef<'a, S>
+    where
+        S: ArgInto<'a>,
+    {
         ParegRef::new(&self.args, Cow::Owned(self.cur.clone()))
+    }
+
+    /// Get the next argument as string.
+    ///
+    /// Note that this may return Some("") if the conversion to string fails.
+    /// (e.g. when converting invalid unicode from OsString).
+    // Iterator impl is not possible because the returned values are borrowed.
+    #[allow(clippy::should_implement_trait)]
+    #[inline]
+    pub fn next_str<'a>(&'a mut self) -> Option<&'a str>
+    where
+        S: ArgInto<'a>,
+    {
+        self.inner().next_str()
     }
 
     /// Get the next argument
     // Iterator impl is not possible because the returned values are borrowed.
     #[allow(clippy::should_implement_trait)]
     #[inline]
-    pub fn next(&mut self) -> Option<&str> {
+    pub fn next<'a>(&'a mut self) -> Option<&'a S>
+    where
+        S: ArgInto<'a>,
+    {
         self.inner().next()
     }
 
     /// Equivalent to calling next `cnt` times.
     #[inline]
-    pub fn skip_args(&mut self, cnt: usize) -> Option<&str> {
+    pub fn skip_args<'a>(&'a mut self, cnt: usize) -> Option<&'a S>
+    where
+        S: ArgInto<'a>,
+    {
         self.inner().skip_args(cnt)
     }
 
     /// Skip all remaining arguments and return the last.
     #[inline]
-    pub fn skip_all(&mut self) -> Option<&str> {
+    pub fn skip_all<'a>(&'a mut self) -> Option<&'a S>
+    where
+        S: ArgInto<'a>,
+    {
         self.inner().skip_all()
     }
 
     /// Jump so that the argument at index `idx` is the next argument. Gets the
     /// argument at `idx - 1`.
     #[inline]
-    pub fn jump(&mut self, idx: usize) -> Option<&str> {
+    pub fn jump<'a>(&'a mut self, idx: usize) -> Option<&'a S>
+    where
+        S: ArgInto<'a>,
+    {
         self.inner().jump(idx)
     }
 
     /// Jump to the zeroth argument.
     #[inline]
-    pub fn reset(&mut self) {
+    pub fn reset<'a>(&'a mut self)
+    where
+        S: ArgInto<'a>,
+    {
         self.inner().reset()
     }
 
     /// Get the last returned argument.
     #[inline]
-    pub fn cur(&self) -> Option<&str> {
+    pub fn cur<'a>(&'a self) -> Option<&'a S>
+    where
+        S: ArgInto<'a>,
+    {
         self.inner().cur()
     }
 
     /// Gets all the arguments (including the first one).
     #[inline]
-    pub fn all_args(&self) -> &[String] {
+    pub fn all_args<'a>(&'a self) -> &'a [S]
+    where
+        S: ArgInto<'a>,
+    {
         self.inner().all_args()
     }
 
     /// Gets the remaining arguments (not including the current).
     #[inline]
-    pub fn remaining(&self) -> &[String] {
+    pub fn remaining<'a>(&'a self) -> &'a [S]
+    where
+        S: ArgInto<'a>,
+    {
         self.inner().remaining()
     }
 
     /// Gets the remaining arguments (including the current).
     #[inline]
-    pub fn cur_remaining(&self) -> &[String] {
+    pub fn cur_remaining<'a>(&'a self) -> &'a [S]
+    where
+        S: ArgInto<'a>,
+    {
         self.inner().cur_remaining()
     }
 
     /// Get value that will be returned with the next call to `next`.
     #[inline]
-    pub fn peek(&self) -> Option<&str> {
+    pub fn peek<'a>(&'a self) -> Option<&'a S>
+    where
+        S: ArgInto<'a>,
+    {
         self.inner().peek()
     }
 
     /// Get the index of the next argument.
     #[inline]
-    pub fn next_idx(&self) -> Option<usize> {
+    pub fn next_idx<'a>(&'a self) -> Option<usize>
+    where
+        S: ArgInto<'a>,
+    {
         self.inner().next_idx()
     }
 
     /// Get index of the current argument.
     #[inline]
-    pub fn cur_idx(&self) -> Option<usize> {
+    pub fn cur_idx<'a>(&'a self) -> Option<usize>
+    where
+        S: ArgInto<'a>,
+    {
         self.inner().cur_idx()
     }
 
     /// Get argument at the given index.
     #[inline]
-    pub fn get(&self, idx: usize) -> Option<&str> {
+    pub fn get<'a>(&'a self, idx: usize) -> Option<&'a S>
+    where
+        S: ArgInto<'a>,
+    {
         self.inner().get(idx)
     }
 
@@ -194,7 +247,10 @@ impl Pareg {
     pub fn next_manual<'a, T: 'a>(
         &'a mut self,
         f: impl Fn(&'a str) -> Result<T>,
-    ) -> Result<T> {
+    ) -> Result<T>
+    where
+        S: ArgInto<'a>,
+    {
         self.inner().next_manual(f)
     }
 
@@ -222,7 +278,10 @@ impl Pareg {
     pub fn cur_manual<'a, T: 'a>(
         &'a self,
         f: impl Fn(&'a str) -> Result<T>,
-    ) -> Result<T> {
+    ) -> Result<T>
+    where
+        S: ArgInto<'a>,
+    {
         self.inner().cur_manual(f)
     }
 
@@ -240,7 +299,10 @@ impl Pareg {
     /// assert_eq!(0.25, args.next_arg::<f64>().unwrap());
     /// ```
     #[inline]
-    pub fn next_arg<'a, T: FromArg<'a>>(&'a mut self) -> Result<T> {
+    pub fn next_arg<'a, T: FromArg<'a>>(&'a mut self) -> Result<T>
+    where
+        S: ArgInto<'a>,
+    {
         self.inner().next_arg()
     }
 
@@ -275,7 +337,10 @@ impl Pareg {
     pub fn next_key_mval<'a, K: FromArg<'a>, V: FromArg<'a>>(
         &'a mut self,
         sep: char,
-    ) -> Result<(K, Option<V>)> {
+    ) -> Result<(K, Option<V>)>
+    where
+        S: ArgInto<'a>,
+    {
         self.inner().next_key_mval(sep)
     }
 
@@ -306,7 +371,10 @@ impl Pareg {
     pub fn next_key_val<'a, K: FromArg<'a>, V: FromArg<'a>>(
         &'a mut self,
         sep: char,
-    ) -> Result<(K, V)> {
+    ) -> Result<(K, V)>
+    where
+        S: ArgInto<'a>,
+    {
         self.inner().next_key_val(sep)
     }
 
@@ -328,7 +396,10 @@ impl Pareg {
     /// assert_eq!(false, args.next_bool("always", "never").unwrap());
     /// ```
     #[inline]
-    pub fn next_bool(&mut self, t: &str, f: &str) -> Result<bool> {
+    pub fn next_bool<'a>(&'a mut self, t: &str, f: &str) -> Result<bool>
+    where
+        S: ArgInto<'a>,
+    {
         self.inner().next_bool(t, f)
     }
 
@@ -359,12 +430,15 @@ impl Pareg {
     /// );
     /// ```
     #[inline]
-    pub fn next_opt_bool(
-        &mut self,
+    pub fn next_opt_bool<'a>(
+        &'a mut self,
         t: &str,
         f: &str,
         n: &str,
-    ) -> Result<Option<bool>> {
+    ) -> Result<Option<bool>>
+    where
+        S: ArgInto<'a>,
+    {
         self.inner().next_opt_bool(t, f, n)
     }
 
@@ -391,7 +465,10 @@ impl Pareg {
     /// );
     /// ```
     #[inline]
-    pub fn next_key<'a, T: FromArg<'a>>(&'a mut self, sep: char) -> Result<T> {
+    pub fn next_key<'a, T: FromArg<'a>>(&'a mut self, sep: char) -> Result<T>
+    where
+        S: ArgInto<'a>,
+    {
         self.inner().next_key(sep)
     }
 
@@ -419,7 +496,10 @@ impl Pareg {
     /// );
     /// ```
     #[inline]
-    pub fn next_val<'a, T: FromArg<'a>>(&'a mut self, sep: char) -> Result<T> {
+    pub fn next_val<'a, T: FromArg<'a>>(&'a mut self, sep: char) -> Result<T>
+    where
+        S: ArgInto<'a>,
+    {
         self.inner().next_val(sep)
     }
 
@@ -454,7 +534,10 @@ impl Pareg {
     pub fn next_mval<'a, T: FromArg<'a>>(
         &'a mut self,
         sep: char,
-    ) -> Result<Option<T>> {
+    ) -> Result<Option<T>>
+    where
+        S: ArgInto<'a>,
+    {
         self.inner().next_mval(sep)
     }
 
@@ -475,7 +558,10 @@ impl Pareg {
     /// assert_eq!(0.25, args.cur_arg::<f64>().unwrap());
     /// ```
     #[inline]
-    pub fn cur_arg<'a, T: FromArg<'a>>(&'a self) -> Result<T> {
+    pub fn cur_arg<'a, T: FromArg<'a>>(&'a self) -> Result<T>
+    where
+        S: ArgInto<'a>,
+    {
         self.inner().cur_arg()
     }
 
@@ -514,7 +600,10 @@ impl Pareg {
     pub fn cur_key_mval<'a, K: FromArg<'a>, V: FromArg<'a>>(
         &'a self,
         sep: char,
-    ) -> Result<(K, Option<V>)> {
+    ) -> Result<(K, Option<V>)>
+    where
+        S: ArgInto<'a>,
+    {
         self.inner().cur_key_mval(sep)
     }
 
@@ -548,7 +637,10 @@ impl Pareg {
     pub fn cur_key_val<'a, K: FromArg<'a>, V: FromArg<'a>>(
         &'a self,
         sep: char,
-    ) -> Result<(K, V)> {
+    ) -> Result<(K, V)>
+    where
+        S: ArgInto<'a>,
+    {
         self.inner().cur_key_val(sep)
     }
 
@@ -574,7 +666,10 @@ impl Pareg {
     /// assert_eq!(false, args.cur_bool("always", "never").unwrap());
     /// ```
     #[inline]
-    pub fn cur_bool(&self, t: &str, f: &str) -> Result<bool> {
+    pub fn cur_bool<'a>(&'a self, t: &str, f: &str) -> Result<bool>
+    where
+        S: ArgInto<'a>,
+    {
         self.inner().cur_bool(t, f)
     }
 
@@ -609,12 +704,15 @@ impl Pareg {
     /// );
     /// ```
     #[inline]
-    pub fn cur_opt_bool(
-        &self,
+    pub fn cur_opt_bool<'a>(
+        &'a self,
         t: &str,
         f: &str,
         n: &str,
-    ) -> Result<Option<bool>> {
+    ) -> Result<Option<bool>>
+    where
+        S: ArgInto<'a>,
+    {
         self.inner().cur_opt_bool(t, f, n)
     }
 
@@ -644,7 +742,10 @@ impl Pareg {
     /// );
     /// ```
     #[inline]
-    pub fn cur_key<'a, T: FromArg<'a>>(&'a self, sep: char) -> Result<T> {
+    pub fn cur_key<'a, T: FromArg<'a>>(&'a self, sep: char) -> Result<T>
+    where
+        S: ArgInto<'a>,
+    {
         self.inner().cur_key(sep)
     }
 
@@ -675,7 +776,10 @@ impl Pareg {
     /// );
     /// ```
     #[inline]
-    pub fn cur_val<'a, T: FromArg<'a>>(&'a self, sep: char) -> Result<T> {
+    pub fn cur_val<'a, T: FromArg<'a>>(&'a self, sep: char) -> Result<T>
+    where
+        S: ArgInto<'a>,
+    {
         self.inner().cur_val(sep)
     }
 
@@ -714,7 +818,10 @@ impl Pareg {
     pub fn cur_mval<'a, T: FromArg<'a>>(
         &'a self,
         sep: char,
-    ) -> Result<Option<T>> {
+    ) -> Result<Option<T>>
+    where
+        S: ArgInto<'a>,
+    {
         self.inner().cur_mval(sep)
     }
 
@@ -744,7 +851,10 @@ impl Pareg {
     pub fn cur_val_or_next<'a, T: FromArg<'a>>(
         &'a mut self,
         sep: char,
-    ) -> Result<T> {
+    ) -> Result<T>
+    where
+        S: ArgInto<'a>,
+    {
         self.inner().cur_val_or_next(sep)
     }
 
@@ -755,7 +865,10 @@ impl Pareg {
         &'a self,
         res: &mut Option<T>,
         f: impl FnOnce(&'a str) -> Result<T>,
-    ) -> Result<()> {
+    ) -> Result<()>
+    where
+        S: ArgInto<'a>,
+    {
         self.inner().try_set_cur_with(res, f)
     }
 
@@ -766,7 +879,10 @@ impl Pareg {
         &'a mut self,
         res: &mut Option<T>,
         f: impl FnOnce(&'a str) -> Result<T>,
-    ) -> Result<()> {
+    ) -> Result<()>
+    where
+        S: ArgInto<'a>,
+    {
         self.inner().try_set_next_with(res, f)
     }
 
@@ -776,7 +892,10 @@ impl Pareg {
     pub fn try_set_cur<'a, T: FromArg<'a>>(
         &'a mut self,
         res: &mut Option<T>,
-    ) -> Result<()> {
+    ) -> Result<()>
+    where
+        S: ArgInto<'a>,
+    {
         self.inner().try_set_cur(res)
     }
 
@@ -786,7 +905,10 @@ impl Pareg {
     pub fn try_set_next<'a, T: FromArg<'a>>(
         &'a mut self,
         res: &mut Option<T>,
-    ) -> Result<()> {
+    ) -> Result<()>
+    where
+        S: ArgInto<'a>,
+    {
         self.inner().try_set_next(res)
     }
 
@@ -796,10 +918,10 @@ impl Pareg {
     /// Difference from [`Pareg::cur_list`] is that this will first to split
     /// and than try to parse.
     #[inline]
-    pub fn split_cur<'a, T: FromArg<'a>>(
-        &'a self,
-        sep: &str,
-    ) -> Result<Vec<T>> {
+    pub fn split_cur<'a, T: FromArg<'a>>(&'a self, sep: &str) -> Result<Vec<T>>
+    where
+        S: ArgInto<'a>,
+    {
         self.inner().split_cur(sep)
     }
 
@@ -810,7 +932,10 @@ impl Pareg {
     /// of `sep`, and it will properly parse the vales, whereas
     /// [`Pareg::split_cur`] would split `arg` and than try to parse.
     #[inline]
-    pub fn cur_list<T: FromRead>(&self, sep: &str) -> Result<Vec<T>> {
+    pub fn cur_list<'a, T: FromRead>(&'a self, sep: &str) -> Result<Vec<T>>
+    where
+        S: ArgInto<'a>,
+    {
         self.inner().cur_list(sep)
     }
 
@@ -823,7 +948,10 @@ impl Pareg {
     pub fn split_next<'a, T: FromArg<'a>>(
         &'a mut self,
         sep: &str,
-    ) -> Result<Vec<T>> {
+    ) -> Result<Vec<T>>
+    where
+        S: ArgInto<'a>,
+    {
         self.inner().split_next(sep)
     }
 
@@ -834,59 +962,92 @@ impl Pareg {
     /// of `sep`, and it will properly parse the vales, whereas
     /// [`Pareg::split_next`] would split `arg` and than try to parse.
     #[inline]
-    pub fn next_list<T: FromRead>(&mut self, sep: &str) -> Result<Vec<T>> {
+    pub fn next_list<'a, T: FromRead>(
+        &'a mut self,
+        sep: &str,
+    ) -> Result<Vec<T>>
+    where
+        S: ArgInto<'a>,
+    {
         self.inner().next_list(sep)
     }
 
     /// Leave parsing of the next arguments to the `FromArgs` implementation of
     /// `T`.
-    pub fn next_sub<'a, T: FromArgs<'a>>(&'a mut self) -> Result<T> {
+    pub fn next_sub<'a, T: FromArgs<'a>>(&'a mut self) -> Result<T>
+    where
+        S: ArgInto<'a>,
+    {
         self.inner().next_sub()
     }
 
     /// Leave the parsing of the current and following arguments to the
     /// `FromArgs` implementation of `T`.
-    pub fn cur_sub<'a, T: FromArgs<'a>>(&'a mut self) -> Result<T> {
+    pub fn cur_sub<'a, T: FromArgs<'a>>(&'a mut self) -> Result<T>
+    where
+        S: ArgInto<'a>,
+    {
         self.inner().cur_sub()
     }
 
     /// Creates pretty error that the last argument (cur) is unknown.
     #[inline]
-    pub fn err_unknown_argument(&self) -> ArgError {
+    pub fn err_unknown_argument<'a>(&'a self) -> ArgError
+    where
+        S: ArgInto<'a>,
+    {
         self.inner().err_unknown_argument()
     }
 
     /// Creates pretty error that there should be more arguments but there are
     /// no more arguments.
     #[inline]
-    pub fn err_no_more_arguments(&self) -> ArgError {
+    pub fn err_no_more_arguments<'a>(&'a self) -> ArgError
+    where
+        S: ArgInto<'a>,
+    {
         self.inner().err_no_more_arguments()
     }
 
     /// Creates error that sais that the current argument is specified too many
     /// times.
     #[inline]
-    pub fn err_cur_too_many_arguments(&self) -> ArgError {
+    pub fn err_cur_too_many_arguments<'a>(&'a self) -> ArgError
+    where
+        S: ArgInto<'a>,
+    {
         self.inner().err_cur_too_many_arguments()
     }
 
     /// Creates error that says that the current argument has invalid value.
     #[inline]
-    pub fn err_invalid(&self) -> ArgError {
+    pub fn err_invalid<'a>(&'a self) -> ArgError
+    where
+        S: ArgInto<'a>,
+    {
         self.inner().err_invalid()
     }
 
     /// Creates error that says that the given part of the current argument has
     /// invalid value.
     #[inline]
-    pub fn err_invalid_value(&self, value: String) -> ArgError {
+    pub fn err_invalid_value<'a>(&'a self, value: String) -> ArgError
+    where
+        S: ArgInto<'a>,
+    {
         self.inner().err_invalid_value(value)
     }
 
     /// Creates error that says that the given part of the current argument has
     /// invalid value.
     #[inline]
-    pub fn err_invalid_span(&self, span: impl RangeBounds<usize>) -> ArgError {
+    pub fn err_invalid_span<'a>(
+        &'a self,
+        span: impl RangeBounds<usize>,
+    ) -> ArgError
+    where
+        S: ArgInto<'a>,
+    {
         self.inner().err_invalid_span(span)
     }
 
@@ -894,7 +1055,10 @@ impl Pareg {
     /// message. Consider using [`ParegRef::cur_manual`] or
     /// [`ParegRef::next_manual`] instead.
     #[inline]
-    pub fn map_err(&self, err: ArgError) -> ArgError {
+    pub fn map_err<'a>(&'a self, err: ArgError) -> ArgError
+    where
+        S: ArgInto<'a>,
+    {
         self.inner().map_err(err)
     }
 
@@ -916,7 +1080,32 @@ impl Pareg {
     /// assert_eq!((10, 0.25), res);
     /// ```
     #[inline]
-    pub fn map_res<T>(&self, res: Result<T>) -> Result<T> {
+    pub fn map_res<'a, T>(&'a self, res: Result<T>) -> Result<T>
+    where
+        S: ArgInto<'a>,
+    {
         self.inner().map_res(res)
+    }
+}
+
+impl Pareg<String> {
+    /// Create [`Pareg`] from [`env::args`], the first argument is skipped.
+    #[inline]
+    pub fn args() -> Self {
+        Self {
+            args: env::args().collect(),
+            cur: 1.into(),
+        }
+    }
+}
+
+impl Pareg<OsString> {
+    /// Create [`Pareg`] from [`env::args`], the first argument is skipped.
+    #[inline]
+    pub fn args_os() -> Self {
+        Self {
+            args: env::args_os().collect(),
+            cur: 1.into(),
+        }
     }
 }